// src/collections/product.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

import {
	defineCollection,
	type GridCommand,
} from "@katren/vue-collection-lib";
import { productApi } from "@/api/product.gen";
import type { ProductList } from "@/types/productList.gen";
import type {
	Product,
	ProductKey,
	ProductNew,
	ProductUpd,
} from "@/types/product.gen";

type ListModel = ProductList;
type CreateModel = ProductNew;
type UpdateModel = ProductUpd;
const commands: GridCommand<ListModel, ProductKey>[] = [
	{ name: "create" },
	{ name: "edit" },
	{ name: "copy" },
	{ name: "delete" },
	{ name: "search" },
	{ name: "refresh" },
];

export const productCollection = defineCollection<
	ListModel,
	ProductKey,
	CreateModel,
	UpdateModel,
	Product
>({
	titleKey: "Product.title",
	api: productApi,
	columns: [
		{
			field: "id",
			headerKey: "Product.fields.id",
			sortable: true,
			dataType: "number",
			align: "right",
			width: "8rem",
		},
		{
			field: "name",
			headerKey: "Product.fields.name",
			sortable: true,
			width: "32rem",
		},
		{
			field: "measure_unit",
			headerKey: "Product.fields.measure_unit",
			sortable: true,
			width: "12rem",
		},
		{
			field: "is_active",
			headerKey: "Product.fields.is_active",
			sortable: true,
			dataType: "boolean",
			width: "8rem",
		},
	],
	commands,
	dataKey: "id",
	getKey: (row) => ({
		id: row.id,
	}),
	routes: {
		create: () => ({ name: "productCreate" }),
		edit: (row) => ({
			name: "productEdit",
			params: {
				id: String(row.id),
			},
		}),
		copy: (row) => ({
			name: "productCreate",
			query: {
				copy_id: String(row.id),
			},
		}),
	},
	editMode: "page",
	stateKey: "product-grid",
	pageSize: 30,
	showCommandShortcuts: false,
});