// src/types/productList.gen.ts
// Code generated by cmd/codegen. DO NOT EDIT.

export interface ProductListDTO {
	id: number;
	parent_id: number | null;
	name: string;
	measure_unit_id: number | null;
	ref_1c: Record<string, unknown> | null;
	measure_unit: Record<string, unknown> | null;
	sort_order: number;
	is_group: boolean;
	is_active: boolean;
}

export interface ProductList {
	id: number;
	parent_id: number | null;
	name: string;
	measure_unit_id: number | null;
	ref_1c: Record<string, unknown> | null;
	measure_unit: Record<string, unknown> | null;
	sort_order: number;
	is_group: boolean;
	is_active: boolean;
}